Added leechblock and optimize extension script

+++ usr/share/system-scripts/update-firefox/extension/cors-everywhere.sh
@@ -1,6 +1,10 @@
-cd /tmp
-wget https://codeload.github.com/spenibus/cors-everywhere-firefox-addon/zip/refs/heads/master
-unzip master
-cd cors-everywhere-firefox-addon-master
-rm -rf _test
-zip "$FIREFOXPROFILEPATH/extensions/cors-everywhere@spenibus.xpi" * media/*
+if [ ! -f "$FIREFOXPROFILEPATH/extensions/cors-everywhere@spenibus.xpi" ]; then
+	cd /tmp
+	if [ ! -d "/tmp/cors-everywhere-firefox-addon-master" ]; then
+		wget --output-document="cors-everywhere.zip" https://codeload.github.com/spenibus/cors-everywhere-firefox-addon/zip/refs/heads/master
+		unzip -q cors-everywhere.zip
+	fi
+	cd /tmp/cors-everywhere-firefox-addon-master
+	rm -rf _test
+	zip -q -r "$FIREFOXPROFILEPATH/extensions/cors-everywhere@spenibus.xpi" *
+fi
